Refactor code, remove cargo settings
